Request permission works, optimized imports, refactor strings.xml
