+++ app/src/main/java/es/rodal/keoapp/KeoApp.kt
@@ -20,7 +20,6 @@ import android.app.Application
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
-import android.content.Intent
 import android.os.Build
 import dagger.hilt.android.HiltAndroidApp
