+++ app/src/main/java/es/rodal/keoapp/MainActivity.kt
@@ -16,31 +16,16 @@
 
 package es.rodal.keoapp
 
-import android.app.AlarmManager
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
-import android.provider.Settings
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import androidx.core.content.ContextCompat
 import dagger.hilt.android.AndroidEntryPoint
-import es.rodal.keoapp.RecordatorioNotificationReceiver.Companion.NOTIFICATION_ID
 import es.rodal.keoapp.ui.MainNavigation
 import es.rodal.keoapp.ui.theme.MyApplicationTheme
-import java.util.Calendar
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -48,27 +33,13 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        // Verificar y solicitar permiso al abrir la aplicación
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            val alarmManager = getSystemService(Context.ALARM_SERVICE) as AlarmManager
-            if (!alarmManager.canScheduleExactAlarms()) {
-                // Redirigir a la configuración para habilitar el permiso
-                val intent = Intent().apply {
-                    action = Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM
-                }
-                startActivity(intent)
-            }
-        }
-
-        //https://developer.android.com/about/versions/14/changes/schedule-exact-alarms?hl=es-419
-
         setContent {
             MyApplicationTheme {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    MainNavigation(activity = this)
+                    MainNavigation(context = this)
                 }
             }
         }
