+++ app/src/main/java/es/rodal/keoapp/data/local/database/RecordatorioEntity.kt
@@ -16,12 +16,8 @@
 
 package es.rodal.keoapp.data.local.database
 
-import androidx.room.Dao
 import androidx.room.Entity
-import androidx.room.Insert
 import androidx.room.PrimaryKey
-import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
 import java.util.Date
 
 @Entity
