+++ app/src/main/java/es/rodal/keoapp/data/local/di/DatabaseModule.kt
@@ -17,12 +17,10 @@
 package es.rodal.keoapp.data.local.di
 
 import android.app.Application
-import android.content.Context
 import androidx.room.Room
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import es.rodal.keoapp.data.domain.repository.RecordatorioRepository
 import es.rodal.keoapp.data.domain.repository.RecordatorioRepositoryImpl
