+++ app/src/main/java/es/rodal/keoapp/ui/recordatorio/home/RecordatorioScreen.kt
@@ -16,13 +16,7 @@
 
 package es.rodal.keoapp.ui.recordatorio.home
 
-import android.Manifest
-import android.app.AlarmManager
 import android.content.Context
-import android.content.Intent
-import android.os.Build
-import android.provider.Settings
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -34,8 +28,6 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
@@ -48,21 +40,16 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarColors
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.core.content.ContextCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.isGranted
-import com.google.accompanist.permissions.rememberPermissionState
 import es.rodal.keoapp.R
+import es.rodal.keoapp.ui.utils.PermissionAlarmDialog
+import es.rodal.keoapp.ui.utils.PermissionDialog
 
 @Composable
 fun RecordatorioScreen(
@@ -70,7 +57,6 @@ fun RecordatorioScreen(
     askNotificationPermission: Boolean,
     askAlarmPermission: Boolean,
     navigateToRecordatorioEntry: () -> Unit,
-    modifier: Modifier = Modifier,
     viewModel: RecordatorioViewModel = hiltViewModel()
 ) {
     PermissionAlarmDialog(
@@ -82,8 +68,7 @@ fun RecordatorioScreen(
     RecordatorioScaffold(
         navController = navController,
         navigateToRecordatorioEntry = navigateToRecordatorioEntry,
-        viewModel = viewModel,
-        modifier = modifier
+        viewModel = viewModel
     )
 
 }
@@ -93,8 +78,7 @@ fun RecordatorioScreen(
 fun RecordatorioScaffold(
     navController: NavController,
     navigateToRecordatorioEntry: () -> Unit,
-    viewModel: RecordatorioViewModel,
-    modifier: Modifier = Modifier
+    viewModel: RecordatorioViewModel
 ) {
     Scaffold(
         topBar = {
@@ -107,7 +91,7 @@ fun RecordatorioScaffold(
                     scrolledContainerColor = MaterialTheme.colorScheme.primaryContainer
                 ),
                 title = {
-                    Text("Top app bar")
+                    Text(stringResource(id = R.string.app_name))
                 }
             )
         },
@@ -158,14 +142,18 @@ fun RecordatorioLazyColumn(viewModel: RecordatorioViewModel, context: Context/*,
                     .fillMaxWidth()
             ) {
                 Row {
-                    Column(modifier = Modifier.padding(8.dp).weight(2f)) {
+                    Column(modifier = Modifier
+                        .padding(8.dp)
+                        .weight(2f)) {
                         Text(text = recordatorio.name, modifier = Modifier.padding(16.dp))
                         Text(
                             text = recordatorio.recordatorioTime.toString(),
                             modifier = Modifier.padding(8.dp)
                         )
                     }
-                    Column(modifier = Modifier.padding(8.dp).weight(1f)) {
+                    Column(modifier = Modifier
+                        .padding(8.dp)
+                        .weight(1f)) {
                         Button(onClick = { viewModel.deleteRecordatorio(context, recordatorio) }) {
                             Text(stringResource(id = R.string.delete))
                         }
@@ -198,99 +186,4 @@ fun EmptyView() {
     }
 }
 
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun PermissionDialog(
-    askNotificationPermission: Boolean
-) {
-    if (askNotificationPermission && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)) {
-        val notificationPermissionState = rememberPermissionState(Manifest.permission.POST_NOTIFICATIONS)
-        if (!notificationPermissionState.status.isGranted) {
-            val openAlertDialog = remember { mutableStateOf(true) }
 
-            when {
-                openAlertDialog.value -> {
-                    AlertDialog(
-                        icon = {
-                            Icon(
-                                imageVector = Icons.Default.Notifications,
-                                contentDescription = stringResource(R.string.notifications)
-                            )
-                        },
-                        title = {
-                            Text(text = stringResource(R.string.notification_permission_required))
-                        },
-                        text = {
-                            Text(text = stringResource(R.string.notification_permission_required_description_message))
-                        },
-                        onDismissRequest = {
-                            openAlertDialog.value = false
-                        },
-                        confirmButton = {
-                            Button(
-                                onClick = {
-                                    notificationPermissionState.launchPermissionRequest()
-                                    openAlertDialog.value = false
-                                }
-                            ) {
-                                Text(stringResource(R.string.allow))
-                            }
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun PermissionAlarmDialog(
-    askAlarmPermission: Boolean
-) {
-    val context = LocalContext.current
-    val alarmManager = ContextCompat.getSystemService(context, AlarmManager::class.java)
-    if (askAlarmPermission && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE)) {
-        val alarmPermissionState = rememberPermissionState(Manifest.permission.SCHEDULE_EXACT_ALARM)
-        if (alarmManager?.canScheduleExactAlarms() == false) {
-            val openAlertDialog = remember { mutableStateOf(true) }
-
-            when {
-                openAlertDialog.value -> {
-
-                    AlertDialog(
-                        icon = {
-                            Icon(
-                                imageVector = Icons.Default.Notifications,
-                                contentDescription = stringResource(R.string.alarms)
-                            )
-                        },
-                        title = {
-                            Text(text = stringResource(R.string.alarms_permission_required))
-                        },
-                        text = {
-                            Text(text = stringResource(R.string.alarms_permission_required_description_message))
-                        },
-                        onDismissRequest = {
-                            openAlertDialog.value = false
-                        },
-                        confirmButton = {
-                            Button(
-                                onClick = {
-                                    Intent().also { intent ->
-                                        intent.action = Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM
-                                        context.startActivity(intent)
-                                    }
-
-                                    openAlertDialog.value = false
-                                }
-                            ) {
-                                Text(stringResource(R.string.allow))
-                            }
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
