+++ app/src/main/java/es/rodal/keoapp/ui/recordatorio/home/RecordatorioViewModel.kt
@@ -16,12 +16,10 @@
 
 package es.rodal.keoapp.ui.recordatorio.home
 
-import android.app.NotificationManager
 import android.content.Context
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.core.content.ContextCompat.getSystemService
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
