+++ app/src/main/java/es/rodal/keoapp/ui/utils/PermissionUtils.kt
@@ -1,0 +1,137 @@
+package es.rodal.keoapp.ui.utils
+
+import android.Manifest
+import android.app.AlarmManager
+import android.content.Context
+import android.content.Context.ALARM_SERVICE
+import android.content.Intent
+import android.os.Build
+import android.provider.Settings
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.core.content.ContextCompat
+import androidx.core.content.ContextCompat.getSystemService
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.isGranted
+import com.google.accompanist.permissions.rememberPermissionState
+import es.rodal.keoapp.R
+
+
+fun scheduleAlarmPermissionGranted(context: Context): Boolean {
+    val alarmManager: AlarmManager = context.getSystemService(ALARM_SERVICE) as AlarmManager
+    return if ((Build.VERSION.SDK_INT >= Build.VERSION_CODES.S
+                && Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU)
+        && !alarmManager.canScheduleExactAlarms()
+    ) {
+        false
+    } else {
+        true
+    }
+}
+
+
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun PermissionDialog(
+    askNotificationPermission: Boolean
+) {
+    if (askNotificationPermission && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)) {
+        val notificationPermissionState = rememberPermissionState(Manifest.permission.POST_NOTIFICATIONS)
+        if (!notificationPermissionState.status.isGranted) {
+            val openAlertDialog = remember { mutableStateOf(true) }
+
+            when {
+                openAlertDialog.value -> {
+                    AlertDialog(
+                        icon = {
+                            Icon(
+                                imageVector = Icons.Default.Notifications,
+                                contentDescription = stringResource(R.string.notifications)
+                            )
+                        },
+                        title = {
+                            Text(text = stringResource(R.string.notification_permission_required))
+                        },
+                        text = {
+                            Text(text = stringResource(R.string.notification_permission_required_description_message))
+                        },
+                        onDismissRequest = {
+                            openAlertDialog.value = false
+                        },
+                        confirmButton = {
+                            Button(
+                                onClick = {
+                                    notificationPermissionState.launchPermissionRequest()
+                                    openAlertDialog.value = false
+                                }
+                            ) {
+                                Text(stringResource(R.string.allow))
+                            }
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
+
+@OptIn(ExperimentalPermissionsApi::class)
+@Composable
+fun PermissionAlarmDialog(
+    askAlarmPermission: Boolean
+) {
+    val context = LocalContext.current
+    val alarmManager = ContextCompat.getSystemService(context, AlarmManager::class.java)
+    if (askAlarmPermission && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE)) {
+        val alarmPermissionState = rememberPermissionState(Manifest.permission.SCHEDULE_EXACT_ALARM)
+        if (alarmManager?.canScheduleExactAlarms() == false) {
+            val openAlertDialog = remember { mutableStateOf(true) }
+
+            when {
+                openAlertDialog.value -> {
+
+                    AlertDialog(
+                        icon = {
+                            Icon(
+                                imageVector = Icons.Default.Notifications,
+                                contentDescription = stringResource(R.string.alarms)
+                            )
+                        },
+                        title = {
+                            Text(text = stringResource(R.string.alarms_permission_required))
+                        },
+                        text = {
+                            Text(text = stringResource(R.string.alarms_permission_required_description_message))
+                        },
+                        onDismissRequest = {
+                            openAlertDialog.value = false
+                        },
+                        confirmButton = {
+                            Button(
+                                onClick = {
+                                    Intent().also { intent ->
+                                        intent.action = Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM
+                                        context.startActivity(intent)
+                                    }
+
+                                    openAlertDialog.value = false
+                                }
+                            ) {
+                                Text(stringResource(R.string.allow))
+                            }
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
