+++ app/src/main/res/values/strings.xml
@@ -38,23 +38,23 @@
     <string name="dosage_and_frequency_mismatch_error_message">Estás seleccionando %s veces de días que son más que la cantidad de dosis.</string>
     <string name="your_plan_for_today">Tu plan para hoy</string>
     <string name="daily_medicine_log">%d de %d completado.</string>
-    <string name="recordatorio_break">Pausa para la medicación</string>
-    <string name="home_screen_empty_card_message">No hay medicamentos programados para esta fecha. Tómate un descanso y relájate.</string>
+    <string name="recordatorio_break">Pausa para la recordar</string>
+    <string name="home_screen_empty_card_message">No hay recordatorios programados para esta fecha. Tómate un descanso y relájate.</string>
     <string name="today">Hoy</string>
     <string name="this_week">Esta semana</string>
     <string name="next_week">La próxima semana</string>
     <string name="notifications">Notificaciones</string>
     <string name="notification_permission_required">Se requiere permiso de notificación</string>
-    <string name="notification_permission_required_description_message">Para asegurarse de que nunca olvide su medicamento, otorgue el permiso de notificación.</string>
+    <string name="notification_permission_required_description_message">Para asegurarse de que nunca olvide su recordaotorio, otorgue el permiso de notificación.</string>
     <string name="allow">Permitir</string>
     <string name="recordatorio_taken_at">Tomado en %s</string>
     <string name="recordatorio_skipped_at">Saltado en %s</string>
     <string name="recordatorio_scheduled_at">Programado a las %s</string>
     <string name="recordatorio_timely_reminders_setup_message">%s ahora está configurado para recibir recordatorios oportunos.</string>
-    <string name="recordatorio_logged">Medicación registrada</string>
+    <string name="recordatorio_logged">Recordatorio registrado</string>
     <string name="recordatorio_dose_details">%d dosis en %s</string>
     <string name="recordatorio_reminder">Recordatorio de medicación</string>
-    <string name="notifications_for_recordatorio_reminder">Notificaciones de recordatorio de medicación</string>
+    <string name="notifications_for_recordatorio_reminder">Notificaciones de recordatorios</string>
     <string name="recordatorio_reminder_time">Es hora de tomar tu medicación, %s. Abra la aplicación para iniciar sesión.</string>
     <string name="cancel">Cancelar</string>
     <string name="ok">DE ACUERDO</string>
